Added MD specific calcfunction
Sometimes the wavefunctions are not salvagable due to some unrecoverable error. In these cases it is important to then do a "dirt" restart, which means starting from scratch but using the atomic positions and velocities of the last MD step. `get_structure_from_trajectory` outputs these quantities in appropriate formats. The positions are simply outputted as a StructureData instance which could then be used as the new input structure. Velocities on the other hand are outputted as Dic instance in `settings` dictionary under `ATOMIC_VELOCITIES` card which is where QE expects velocities. It is important to also update `IONS` card with kywords {'ion_velocities': 'from_input'}. But this needs to be done inside the WorkChain.