Added MD specific calcfunction
In some cases one might want to start an FPMD from a trajectory that was not generated with full first-principles, but can still be used as a good equilibration step. In this case it would be prudent to not combine this trajectory with FPMD trajectories but the atomic positions and velocities may serve as a good starting point for running FPMD. So this `get_last_step_from_trajectory` function is subtley different from `get_structure_from_trajectory`, since it outputs a single frame of a given MD trajectory as a TrajectoryData instance.