Now correctly parses MD calculation
The PwBaseWorkChain was able to launch an MD or variable cell MD calculation but presumably no one ever tested it as the parsing was completely broken. Now it can correctly parse the `.out` file to generate TrajectoryData which contains 'cells',
'steps', 'energy', 'forces', 'stress', 'positions', 'velocities', 'total_force', 'scf_accuracy', 'scf_iterations', and 'energy_accuracy' as arrays, and simulation time and units or forces and velocities as attributes.

All of these changes to parsing are only triggered if the calculation type is `md` or `vc-md` ensuring that for non-MD calculation the functionality of the underlying class is exactly as before.

In addition, there is another optional change which is currently commented out that changes the parsing keyword of `.out` file if the QE calculation is run on SIRIUS enabled platform as their parsing style is different to vanilla QE. So this keyword needs to be enabled if one is running relax type calculations. The keyword for MD calculation remains the same and needs not be changed.